Added styling to style.css per instructions.
When I added the styling, I did not separate the
child selector from the parent selector. This
didn't work, of course.
A dot operator is necessary between selectors when
the selectors are in the same element.
When you are using parent child elements, a space
is required between the parent and child selectors.